完成资讯百分之80

+++ BoBanTang/BBTCampusInfoManager.h
@@ -13,7 +13,7 @@
 
 @property (strong, nonatomic) NSMutableArray * infoArray;                   //Stores a list of infos
 @property (strong, nonatomic) NSMutableArray * collectedInfoArray;          //This array stores info with all properties
-@property (assign, nonatomic) int infoCount;                                //Count the number of infos that have been loaded in total
+@property (assign, nonatomic) unsigned long infoCount;                                //Count the number of infos that have been loaded in total
 
 + (instancetype)sharedInfoManager;                                          //Singleton method
 - (void)loadMoreData;
